添加admin后台UI页面，添加控制器

+++ shop-admin/src/main/java/com/zhuofengyuan/wechat/shop/admin/kengen/KeyGeneratorController.java
@@ -23,7 +23,7 @@ public class KeyGeneratorController {
 
     @RequestMapping("/keygen")
     public String generateKey() {
-        return String.valueOf(keyGenerator.generateKey().longValue());
+        return String.valueOf(keyGenerator.generateKey());
     }
 
 }
